Convert account, admin, and auth routes to SSR with shared loaders and embedded loader data; add /reset-password HTML route
