+++ packages/worker/src/app/handlers/account-integrations.ts
@@ -1,33 +1,10 @@
 import { type Action } from 'remix/router'
+import { loadAccountIntegrationsData } from '#app/account-integrations-data.ts'
 import { redirectToLogin } from '#app/auth-redirect.ts'
 import { readAuthSessionResult } from '#app/auth-session.ts'
 import { readAuthenticatedAppUser } from '#app/authenticated-user.ts'
 import { renderAppPage } from '#app/ssr-render.tsx'
 import { type routes } from '#app/routes.ts'
-import {
-	type IntegrationConfig,
-	parseIntegrationConfig,
-	parseIntegrationJson,
-	parseIntegrationValueName,
-} from '#mcp/capabilities/values/integration-shared.ts'
-import { listValues } from '#mcp/values/service.ts'
-
-type AuthenticatedUser = NonNullable<
-	Awaited<ReturnType<typeof readAuthenticatedAppUser>>
->
-
-type AccountIntegrationListItem = IntegrationConfig & {
-	valueName: string
-	createdAt: string
-	updatedAt: string
-}
-
-type AccountIntegrationsPayload = {
-	ok: true
-	email: string
-	username: string
-	integrations: Array<AccountIntegrationListItem>
-}
 
 export function createAccountIntegrationsHandler(env: Env) {
 	return {
@@ -38,10 +15,17 @@ export function createAccountIntegrationsHandler(env: Env) {
 				return redirectToLogin(request)
 			}
 
+			const user = await readAuthenticatedAppUser(request, env)
+			if (!user) {
+				return redirectToLogin(request)
+			}
+
+			const accountIntegrations = await loadAccountIntegrationsData(env, user)
 			return renderAppPage({
 				request,
 				env,
 				title: 'Integrations',
+				loaderData: { accountIntegrations },
 			})
 		},
 	} satisfies Action<typeof routes.accountIntegrations>
@@ -60,46 +44,9 @@ export function createAccountIntegrationsApiHandler(env: Env) {
 				return jsonResponse({ ok: false, error: 'Method not allowed.' }, 405)
 			}
 
-			return jsonResponse(await buildPayload({ env, user }))
+			return jsonResponse(await loadAccountIntegrationsData(env, user))
 		},
 	} satisfies Action<typeof routes.accountIntegrationsApi>
-}
-
-async function buildPayload(input: {
-	env: Env
-	user: AuthenticatedUser
-}): Promise<AccountIntegrationsPayload> {
-	const values = await listValues({
-		env: input.env,
-		userId: input.user.mcpUser.userId,
-		scope: 'user',
-		storageContext: { sessionId: null, appId: null },
-	})
-	const integrations = values
-		.map((value) => {
-			const integrationName = parseIntegrationValueName(value.name)
-			if (!integrationName) return null
-			const config = parseIntegrationConfig(
-				parseIntegrationJson(value.value),
-				integrationName,
-			)
-			if (!config) return null
-			return {
-				...config,
-				valueName: value.name,
-				createdAt: value.createdAt,
-				updatedAt: value.updatedAt,
-			}
-		})
-		.filter((entry): entry is AccountIntegrationListItem => Boolean(entry))
-		.sort((left, right) => left.name.localeCompare(right.name))
-
-	return {
-		ok: true,
-		email: input.user.email,
-		username: input.user.username,
-		integrations,
-	}
 }
 
 function jsonResponse(body: Record<string, unknown>, status = 200) {
