+++ packages/worker/src/app/handlers/reset-password.ts
@@ -1,0 +1,16 @@
+import { type Action } from 'remix/router'
+import { renderAppPage } from '#app/ssr-render.tsx'
+import { type routes } from '#app/routes.ts'
+
+export function createResetPasswordHandler(env: Env) {
+	return {
+		middleware: [],
+		async handler({ request }) {
+			return renderAppPage({
+				request,
+				env,
+				title: 'Reset password',
+			})
+		},
+	} satisfies Action<typeof routes.resetPassword>
+}
